tests: split test_trajectory.py into focused modules (fixes #211)
