+++ tests/test_trajectory_optimization.py
@@ -1,0 +1,333 @@
+"""Tests for the trajectory optimiser.
+
+Covers: spline construction, cost sub-terms, via-point support,
+optimisation convergence, COM constraint enforcement, parallel
+multi-start, cancellation, stall detection, and solution caching.
+"""
+
+from __future__ import annotations
+
+import threading
+
+import numpy as np
+import pytest
+
+from movement_optimizer.models import (
+    BodyModel,
+    make_squat_config,
+)
+from movement_optimizer.trajectory import (
+    CancelledError,
+    OptimizationResult,
+    ProgressReport,
+    SolutionCache,
+    TrajectoryOptimizer,
+)
+
+# ==============================================================
+# Optimization Tests
+# ==============================================================
+
+
+class TestOptimization:
+    def test_optimize_returns_result(self, squat_optimizer) -> None:
+        opt, _, _, _, _ = squat_optimizer
+        result = opt.optimize()
+        assert isinstance(result, OptimizationResult)
+
+    def test_result_shapes(self, squat_optimizer) -> None:
+        opt, _, _, _, _ = squat_optimizer
+        result = opt.optimize()
+        n = opt.n_eval
+        assert result.t.shape == (n,)
+        assert result.q.shape == (n, 3)
+        assert result.qd.shape == (n, 3)
+        assert result.torques.shape == (n, 3)
+        assert result.power.shape == (n, 3)
+        assert result.com.shape == (n, 2)
+        assert result.bar.shape == (n, 2)
+
+    def test_precondition_objective_finite(self, squat_optimizer) -> None:
+        opt, _, _, _, _ = squat_optimizer
+        wp = opt._initial_guess()
+        cost = opt._compute_cost(wp.flatten())
+        assert cost < float("inf"), "Precondition violated: initial objective is not finite"
+
+    def test_postcondition_kkt_within_tol(self, squat_optimizer) -> None:
+        opt, _, _, _, _ = squat_optimizer
+        # We assume the optimization result includes 'success' which means KKT conditions are within tolerance
+        result = opt.optimize()
+        assert result.success, (
+            "Postcondition violated: optimization did not satisfy KKT within tolerance"
+        )
+
+    def test_cost_decreases(self) -> None:
+        """With enough waypoints, optimization should reduce cost."""
+        body = BodyModel(75.0, 1.75)
+        dyn, qs, qe, qb = make_squat_config(body, 60.0)
+        opt = TrajectoryOptimizer(
+            body,
+            dyn,
+            "squat",
+            60.0,
+            qs,
+            qe,
+            qb,
+            duration=2.0,
+            n_waypoints=12,
+            n_eval=40,
+            n_starts=1,
+        )
+        wp0 = opt._initial_guess()
+        initial_cost = opt._compute_cost(wp0.flatten())
+        result = opt.optimize()
+        assert result.cost <= initial_cost
+
+    def test_horizontal_range_positive(self, squat_optimizer) -> None:
+        opt, _, _, _, _ = squat_optimizer
+        result = opt.optimize()
+        assert result.com_horizontal_range_cm >= 0
+
+    def test_full_squat_reaches_depth(self, full_squat_optimizer) -> None:
+        opt, _, _ = full_squat_optimizer
+        result = opt.optimize()
+        min_thigh_angle = np.min(result.q[:, 1])
+        assert min_thigh_angle < np.radians(-45)
+
+    def test_optimization_succeeds(self) -> None:
+        """Optimization with production-size config should succeed."""
+        body = BodyModel(75.0, 1.75)
+        dyn, qs, qe, qb = make_squat_config(body, 60.0)
+        opt = TrajectoryOptimizer(
+            body,
+            dyn,
+            "squat",
+            60.0,
+            qs,
+            qe,
+            qb,
+            duration=2.0,
+            n_waypoints=12,
+            n_eval=40,
+            n_starts=1,
+        )
+        result = opt.optimize()
+        assert result.cost < float("inf")
+        assert result.success
+
+    def test_com_stays_in_inner_bos(self) -> None:
+        """After SLSQP optimization, COM should stay within the inner 60% BOS."""
+        body = BodyModel(75.0, 1.75)
+        dyn, qs, qe, qb = make_squat_config(body, 60.0)
+        opt = TrajectoryOptimizer(
+            body,
+            dyn,
+            "squat",
+            60.0,
+            qs,
+            qe,
+            qb,
+            duration=2.0,
+            n_waypoints=12,
+            n_eval=40,
+            n_starts=2,
+        )
+        result = opt.optimize()
+        com_x = result.com[:, 0]
+        assert np.all(com_x >= body.inner_heel - 0.01), (
+            f"COM below inner_heel: min={com_x.min():.4f}, bound={body.inner_heel:.4f}"
+        )
+        assert np.all(com_x <= body.inner_toe + 0.01), (
+            f"COM above inner_toe: max={com_x.max():.4f}, bound={body.inner_toe:.4f}"
+        )
+        assert result.success, "Optimization should report success with COM in bounds"
+
+
+# ==============================================================
+# Multi-start
+# ==============================================================
+
+
+class TestMultiStart:
+    def test_perturbed_guess_seed_0_is_baseline(self, squat_optimizer) -> None:
+        opt, _, _, _, _ = squat_optimizer
+        baseline = opt._initial_guess()
+        perturbed = opt._perturbed_guess(0)
+        np.testing.assert_array_equal(baseline, perturbed)
+
+    def test_perturbed_guess_different_seeds_differ(self, squat_optimizer) -> None:
+        opt, _, _, _, _ = squat_optimizer
+        g1 = opt._perturbed_guess(1)
+        g2 = opt._perturbed_guess(2)
+        assert not np.allclose(g1, g2)
+
+    def test_perturbed_guess_within_bounds(self, squat_optimizer) -> None:
+        opt, _, _, _, _ = squat_optimizer
+        for seed in range(10):
+            wp = opt._perturbed_guess(seed)
+            for j in range(3):
+                assert np.all(wp[:, j] >= opt.q_bounds[j, 0])
+                assert np.all(wp[:, j] <= opt.q_bounds[j, 1])
+
+    def test_parallel_multistart_runs(self) -> None:
+        """Multi-start with n_starts > 1 should run and return a result."""
+        body = BodyModel(75.0, 1.75)
+        dyn, qs, qe, qb = make_squat_config(body, 60.0)
+        opt = TrajectoryOptimizer(
+            body,
+            dyn,
+            "squat",
+            60.0,
+            qs,
+            qe,
+            qb,
+            duration=2.0,
+            n_waypoints=6,
+            n_eval=20,
+            n_starts=3,
+        )
+        result = opt.optimize()
+        assert isinstance(result, OptimizationResult)
+        assert result.cost < float("inf")
+
+
+# ==============================================================
+# Cancellation
+# ==============================================================
+
+
+class TestCancellation:
+    def test_cancel_returns_inf(self, squat_optimizer) -> None:
+        opt, _, _, _, _ = squat_optimizer
+        opt.cancel_event.set()
+        wp = opt._initial_guess()
+        result = opt._compute_cost(wp.flatten())
+        assert result == float("inf")
+
+    def test_cancel_during_optimize(self) -> None:
+        """Setting cancel_event before optimize should raise CancelledError."""
+        body = BodyModel(75.0, 1.75)
+        dyn, qs, qe, qb = make_squat_config(body, 60.0)
+        cancel = threading.Event()
+        opt = TrajectoryOptimizer(
+            body,
+            dyn,
+            "squat",
+            60.0,
+            qs,
+            qe,
+            qb,
+            n_waypoints=6,
+            n_eval=20,
+            n_starts=1,
+            cancel_event=cancel,
+        )
+        # Set cancel immediately — should abort on first cost eval or callback
+        cancel.set()
+        with pytest.raises(CancelledError):
+            opt.optimize()
+
+    def test_cancel_event_default(self, squat_optimizer) -> None:
+        opt, _, _, _, _ = squat_optimizer
+        assert not opt.cancel_event.is_set()
+
+
+# ==============================================================
+# Progress Reporting & Stall Detection
+# ==============================================================
+
+
+class TestProgressReporting:
+    def test_progress_callback_receives_report(self) -> None:
+        body = BodyModel(75.0, 1.75)
+        dyn, qs, qe, qb = make_squat_config(body, 60.0)
+        reports: list[ProgressReport] = []
+        # Use larger problem so SLSQP does enough evals to trigger progress
+        opt = TrajectoryOptimizer(
+            body,
+            dyn,
+            "squat",
+            60.0,
+            qs,
+            qe,
+            qb,
+            n_waypoints=12,
+            n_eval=40,
+            n_starts=1,
+            progress_cb=lambda r: reports.append(r),
+        )
+        opt.optimize()
+        # SLSQP may converge fast with hard constraints; reports may be empty
+        # The key test is that optimization completes without error
+        if len(reports) > 0:
+            assert isinstance(reports[0], ProgressReport)
+
+    def test_stall_detection_flat_cost(self, squat_optimizer) -> None:
+        opt, _, _, _, _ = squat_optimizer
+        opt._cost_history = [100.0] * 100
+        is_stalled, _reason = opt._detect_stall()
+        assert is_stalled
+
+    def test_stall_detection_improving(self, squat_optimizer) -> None:
+        opt, _, _, _, _ = squat_optimizer
+        opt._cost_history = list(np.linspace(1000, 100, 100))
+        is_stalled, _ = opt._detect_stall()
+        assert not is_stalled
+
+
+# ==============================================================
+# Solution Cache
+# ==============================================================
+
+
+class TestSolutionCache:
+    def test_cache_miss_returns_none(self) -> None:
+        cache = SolutionCache()
+        result = cache.get("squat", 75.0, 1.75, {"lower_leg": 1.0}, 60.0, 2.0, 1.0)
+        assert result is None
+
+    def test_cache_hit_returns_result(self) -> None:
+        cache = SolutionCache()
+        n = 10
+        dummy = OptimizationResult(
+            t=np.zeros(n),
+            q=np.zeros((n, 3)),
+            qd=np.zeros((n, 3)),
+            qdd=np.zeros((n, 3)),
+            torques=np.zeros((n, 3)),
+            power=np.zeros((n, 3)),
+            com=np.zeros((n, 2)),
+            bar=np.zeros((n, 2)),
+            success=True,
+            cost=42.0,
+            com_horizontal_range_cm=1.5,
+        )
+        mults = {"lower_leg": 1.0, "upper_leg": 1.0, "torso": 1.0}
+        cache.put("squat", 75.0, 1.75, mults, 60.0, 2.0, 1.0, dummy)
+        hit = cache.get("squat", 75.0, 1.75, mults, 60.0, 2.0, 1.0)
+        assert hit is not None
+        assert hit.cost == 42.0
+
+    def test_cache_clear(self) -> None:
+        cache = SolutionCache()
+        n = 10
+        dummy = OptimizationResult(
+            t=np.zeros(n),
+            q=np.zeros((n, 3)),
+            qd=np.zeros((n, 3)),
+            qdd=np.zeros((n, 3)),
+            torques=np.zeros((n, 3)),
+            power=np.zeros((n, 3)),
+            com=np.zeros((n, 2)),
+            bar=np.zeros((n, 2)),
+            success=True,
+            cost=42.0,
+            com_horizontal_range_cm=1.5,
+        )
+        mults = {"lower_leg": 1.0, "upper_leg": 1.0, "torso": 1.0}
+        cache.put("squat", 75.0, 1.75, mults, 60.0, 2.0, 1.0, dummy)
+        cache.clear()
+        assert cache.get("squat", 75.0, 1.75, mults, 60.0, 2.0, 1.0) is None
+
+
+# ==============================================================
